oggzinfo: Fix calculation of content duration.
Previously, oggz_seek(oggz, SEEK_END)  was used to seek to the start of the
last packet. However, for packets which span multiple pages (such as huge
theora frames), you would then need to scan forward to find its granulepos.
We could have done this, but here we take advantage of the fact that we are
already at the end of the file after pass1, so we can just call
oggz_tell_units(oggz) to get the granulepos of the last packet.

Closes ticket:117


git-svn-id: http://svn.annodex.net/liboggz/trunk@1229 8158c8cd-e7e1-0310-9fa4-c5954c97daef