Add tests for oggz_comment_generate, including reference packets as header.  Possibly too many, cull if not needed.
Putting the packets in a header is not ideal, but it allows the use of fixed size arrays (unlike a separate source file, less prone to errors than using structs for this) and avoids cluttering the source.  I doubt the new comment-test.h is subject to copyright (data listing), but as much of the license from comment-test.c should be added if necessary.


git-svn-id: http://svn.annodex.net/liboggz/trunk@2593 8158c8cd-e7e1-0310-9fa4-c5954c97daef