oggz-chop: Fix bug in update of page accumulator, reported by j^ http://trac.annodex.net/ticket/404
This was triggered in instances where all pages in the page accumulator
were to be kept, eg. if all pages had no granulepos marked. The resulting
attempt to shift the contents of the page accumulator by an offset of 0
caused that table's indices to become invalid, which led to a NULL
dereference a few iterations later.

Closes ticket:404


git-svn-id: http://svn.annodex.net/liboggz/trunk@3610 8158c8cd-e7e1-0310-9fa4-c5954c97daef