Update dirac_parse_info for interlace spec changes.
After a patch by David Flynn.

Whether video is coded in frames or fields is set by the source_sampling 
parameter in the video_format presets defined in table C.1, This can be 
overridden by a custom value later in the sequence header. The field 
order for interlaced video is defined by the preset only and cannot be 
overridden.

Also normalize commas at the end of initializers. 


git-svn-id: http://svn.annodex.net/liboggz/trunk@3762 8158c8cd-e7e1-0310-9fa4-c5954c97daef