oggz_read: return an error when a hole (ie. missing sequence number) is detected in the headers of a track, as such header corruption cannot be tolerated by decoders. Beyond the headers, skip holes in data as before for robustness.
This should fix Mozilla bug 463756: https://bugzilla.mozilla.org/show_bug.cgi?id=463756
which crashed in vorbis_synthesize() after attempting to decode with corrupt headers.
The method of this fix is adapted from libvorbisfile, but here should work for any
content type. Tested with the file attached to the above bug, with fishsound-info
and oggplay-info, both of which previously crashed.


git-svn-id: http://svn.annodex.net/liboggz/trunk@3773 8158c8cd-e7e1-0310-9fa4-c5954c97daef