oggz-chop: set attachment if URL has "&download"
This patch adds the HTTP response header:
Content-Disposition: attahment; filename=myvideo.ogg
if the CGI parameter &download exists. This allows a site to
use URLs of the form video.ogg?t=10/20&download to provide
links for the user to directly download a video rather than
play it back in the browser.