Fix an issue with the symlinks for pre-0.9.9 tool names
When files with pre-0.9.9 tool names exist in $(exec_prefix)/bin,
"make DESTDIR=${staging_dir} install" tries to delete them, which
may cause a failure for lack of permissions (quite likely to happen
when building a package as a normal user), and is wrong anyway
(nothing outside of DESTDIR should ever be changed by "make install").

With this patch, it will now cd to "$(DESTDIR)$(bindir)" and create
relative symlinks there, if need be. This preverse the old behavior
when DESTDIR=="", except for the fact the symlinks will be relative
rather than absolute.